SVR Model for Bitcoin Pricing Prediction
This project aims to develop a price prediction model for Bitcoin using SVR. It uses historical data for market price for Bitcoin as well as for a number of financial and Bitcoin specific metrics. Support Vector Regression (SVR) is closely related to Support Vector Machine (SVM); both are supervised learning methods developed from the same foundational principles. SVM is primarily used for classification tasks, whereas SVR is adapted for regression.